Require colon for scheme dispatch in resolver
Before this change the base resolver would dispatch to the resolver
registered for a scheme, if a uri started with the scheme.

Now a colon after the scheme is required to successfully dispatch the
scheme.

E.g. if a resolver is registered for "custom", then the uri
"custom:path/to/artifact" would dispatch to that resolver, but the uri
"custom" would dispatch to the default resolver.

Signed-off-by: Lukas Puehringer <lukas.puehringer@nyu.edu>