runlib: wrap Popen in with block to maybe fix #547
This is a trial-and-error attempt to fix race condition in #547. Using
the with statement here makes Popen close the file handles we pass to it
on exit.

Given that the file handles we pass have been previously opened by us
and in a with block too, I would assume that closing is taken care of
already.

But since the race condition is clearly related to file resources not
being freed (on Windows), this seems worth a try.

Signed-off-by: Lukas Puehringer <lukas.puehringer@nyu.edu>