runlib: retry file removal to workaround #547
As documented in #547 Windows regularly fails to remove a temporary file
used in a subprocess, and so has every attempt to debug the issue.

Since this issue occurs at least once on almost every
CI run, I propose this hacky workaround.

I don't like it, but if it makes our CI not fail all the time, I can
live with it.

Alternatively, I suggest to disable the feature on Windows.

Signed-off-by: Lukas Puehringer <lukas.puehringer@nyu.edu>