runlib: add signer kwarg to run and record methods
Adds optional signer (`securesystemslib.signer.Signer`) arg to runlib's
run/record functions, as alternative way of signing resulting link
metadata, instead of using signing_key, gpg_keyid, or use_default_gpg.

A deprecation warning is added for `signing_key` only, because `signer`
can be used as backwards compatible replacement, and the patch is part
of a series of patches to prepare for the planned removal of
securesystemslib legacy modules `interface` and `keys`, needed by
`signing_key`.

gpg related arguments are not yet deprecated, as the related
implementation in the securesystemslib Signer API is not compatible
(it uses a different public key metadata format).

Note: The patch aims to be minimally invasive, and thus barely refactors
any of the existing signing argument handling in the relevant functions.
Although, it was tempting to simplify the code, it turned out harder
than thought, and therefor not worth the effort, given that these
arguments are bound to deprecation.

Closes #532

Signed-off-by: Lukas Puehringer <lukas.puehringer@nyu.edu>