Refactor setup and signing verifylib tests
Verifylib test setup includes loading keys and signing metadata. This
patch refactors tests to use newly added test CryptoSigners and related
keys, instead of the legacy test files and deprecated `Metablock.sign()`
interface.

It does not change the behavior of the tests, whose subject are the
verifylib verification routines.

This is part of a series of patches to prepare for removal of legacy
securesystemslib interfaces and key file formats.

Signed-off-by: Lukas Puehringer <lukas.puehringer@nyu.edu>