build: add cd GitHub workflow
Add GitHub workflow to build securesystemslib and release on GitHub and
PyPI. The workflow is copied from python-tuf with minimal changes, see:

```
diff \
  <(curl -s https://raw.githubusercontent.com/theupdateframework/python-tuf/a1a2f2bcbec50baaddc5afb529a1e2381e8ae219/.github/workflows/cd.yml) \
  .github/workflows/cd.yml
```

Prerequisites and usage details are described in
https://github.com/theupdateframework/python-tuf/blob/v3.1.0/docs/RELEASE.md.

NOTE: the verify_release script, mentioned in the python-tuf docs, does
not exist for in-toto. But the same script can be used for in-toto, by
changing a few hardcoded infos about the repo in the script header.

Signed-off-by: Lukas Puehringer <lukas.puehringer@nyu.edu>